Move Ctrl and Shift to periphery.
A lot of the commands I type daily involve the Ctrl and Shift keys.
Having them as dual keys was really slowing me down, I observed that I
was typing faster on my Macbook so it was time for a change. I simply
moved the shifts back to their normal positions, Ctrl was moved to
either side of the home row.